[Chore] after delay 화면 전환 방법 수정 #25
