+++ ZOOC/ZOOC/Application/SceneDelegate.swift
@@ -15,7 +15,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-        window?.rootViewController = ZoocTabBarController()
+        window?.rootViewController = UINavigationController(rootViewController: OnboardingLoginViewController())
         self.window?.backgroundColor = .white
         window?.makeKeyAndVisible()
     }
