+++ ZOOC/ZOOC/Presentation/Onboarding/Controller/OnboardingWelcomeViewController.swift
@@ -26,25 +26,28 @@ final class OnboardingWelcomeViewController: UIViewController{
         super.viewDidLoad()
         
         register()
-    }
-    
-    override func viewDidAppear(_ animated: Bool) {
-        super.viewDidAppear(false)
-        
-        pushToSecondaryWelcomeView()
+        updateWelcomeView()
     }
     
     //MARK: - Custom Method
     
     private func register() {
         onboardingWelcomeView.backButton.addTarget(self, action: #selector(backButtonDidTap), for: .touchUpInside)
+        onboardingWelcomeView.nextButton.addTarget(self, action: #selector(nextButtonDidTap), for: .touchUpInside)
     }
     
-    private func pushToSecondaryWelcomeView() {
+    private func updateWelcomeView() {
         DispatchQueue.main.asyncAfter(deadline: DispatchTime.now() + 2) {
-            let secondaryWelcomeViewController = OnboardingWelcomeSecondaryViewController()
-            self.navigationController?.pushViewController(secondaryWelcomeViewController, animated: true)
+            self.onboardingWelcomeView.welcomeLabel.textColor = .zoocGray1
+            self.onboardingWelcomeView.welcomeSubLabel.isHidden = false
+            self.onboardingWelcomeView.welcomeImage.isHidden = false
+            self.onboardingWelcomeView.nextButton.isHidden = false
         }
+    }
+    
+    private func pushToChooseFamilyRoleView() {
+        let onboardingChooseFamilyRoleViewController = OnboardingChooseFamilyRoleViewController()
+        self.navigationController?.pushViewController(onboardingChooseFamilyRoleViewController, animated: true)
     }
     
     //MARK: - Action Method
@@ -52,6 +55,11 @@ final class OnboardingWelcomeViewController: UIViewController{
     @objc
     private func backButtonDidTap() {
         self.navigationController?.popViewController(animated: true)
+    }
+    
+    @objc
+    func nextButtonDidTap() {
+        pushToChooseFamilyRoleView()
     }
 }
