+++ ZOOC/ZOOC/Presentation/Onboarding/View/OnboardingCompleteProfileView.swift
@@ -18,12 +18,48 @@ final class OnboardingCompleteProfileView: UIView {
         $0.setImage(Image.back, for: .normal)
     }
     
-    private var completeProfileLabel = UILabel().then {
+    public var completeProfileLabel = UILabel().then {
         $0.text = "멋진 프로필이네요! \n이제 가족과 함께해보세요"
         $0.textColor = .zoocDarkGray1
         $0.textAlignment = .left
         $0.font = .zoocDisplay1
         $0.numberOfLines = 2
+    }
+    
+    public var completeProfileSubLabel = UILabel().then {
+        $0.text = "가족 코드를 받았나요?"
+        $0.textColor = .zoocDarkGray1
+        $0.textAlignment = .left
+        $0.font = .zoocDisplay1
+        $0.asColor(targetString: "가족 코드", color: .zoocGradientGreen)
+        $0.isHidden = true
+    }
+    
+    public var completeImage = UIImageView().then {
+        $0.image = Image.graphics2
+        $0.isHidden = true
+    }
+    
+    public var getCodeButton = UIButton().then {
+        $0.setTitle("코드를 받았어요", for: .normal)
+        $0.setTitleColor(.white, for: .normal)
+        $0.titleLabel?.font = .zoocSubhead1
+        $0.titleLabel?.textAlignment = .center
+        $0.layer.cornerRadius = 27
+        $0.clipsToBounds = true
+        $0.backgroundColor = .zoocGradientGreen
+        $0.isHidden = true
+    }
+    
+    public var notGetCodeButton = UIButton().then {
+        $0.setTitle("아니요", for: .normal)
+        $0.setTitleColor(.white, for: .normal)
+        $0.titleLabel?.font = .zoocSubhead1
+        $0.titleLabel?.textAlignment = .center
+        $0.layer.cornerRadius = 27
+        $0.clipsToBounds = true
+        $0.backgroundColor = .zoocGradientGreen
+        $0.isHidden = true
     }
     
     //MARK: - Life Cycles
@@ -46,7 +82,7 @@ final class OnboardingCompleteProfileView: UIView {
     }
     
     private func setLayout() {
-        addSubviews(backButton, completeProfileLabel)
+        addSubviews(backButton, completeProfileLabel, completeProfileSubLabel, completeImage, getCodeButton, notGetCodeButton)
         
         backButton.snp.makeConstraints {
             $0.top.equalTo(self.safeAreaLayoutGuide).offset(5)
@@ -58,8 +94,35 @@ final class OnboardingCompleteProfileView: UIView {
             $0.top.equalTo(self.backButton.snp.bottom).offset(56)
             $0.leading.equalToSuperview().offset(30)
         }
+        
+        completeProfileSubLabel.snp.makeConstraints {
+            $0.top.equalTo(self.completeProfileLabel.snp.bottom).offset(20)
+            $0.leading.equalToSuperview().offset(30)
+        }
+        
+        completeImage.snp.makeConstraints {
+            $0.top.equalTo(self.completeProfileSubLabel.snp.bottom).offset(52)
+            $0.centerX.equalToSuperview()
+            $0.width.equalTo(335)
+            $0.height.equalTo(264)
+        }
+        
+        getCodeButton.snp.makeConstraints {
+            $0.top.equalTo(self.completeImage.snp.bottom).offset(68)
+            $0.centerX.equalToSuperview()
+            $0.width.equalTo(315)
+            $0.height.equalTo(50)
+        }
+        
+        notGetCodeButton.snp.makeConstraints {
+            $0.top.equalTo(self.getCodeButton.snp.bottom).offset(10)
+            $0.centerX.equalToSuperview()
+            $0.width.equalTo(315)
+            $0.height.equalTo(50)
+        }
     }
 }
+
